[secmonkey] Replace ELBAuditor DB query with support watcher
Why is this change necessary?
DB query will perform a very inefficient table scan on the item table
which could slow down DB performance when the item table gets large. It
is also not the best solution because the security group may not be in
the database.

This change addresses the need by:
A better solution would be to add the security group results as a
support watcher and find the security group out of the in memory list.

Potential Side Effects:
None